requirements.txt
boto3==1.28.64
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.4.2
